Move queue logic into a submodule
